Add a new extension for ThinBridge (Microsoft Edge)
This adds a WebExtension addon for ThinBridge that allows to
redirect URLs from Edge to other browsers.

The folder "webextension/host" contains a native messaging host
program for testing. If you want to do a quick testing, you can do:

    $ cd webextension/host/
    $ nmake
    $ nmake install

After this setup, you should be able to confirm that Edge redirect
URLs transparently to IE.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>