Edge addon: Move to blank page on redirect from sub frames
We observed that CANCEL_REQUEST from sub frames causes unexpected moving
back on parent main frame. To avoid it, just move to a blank page at sub
frames.

We don't update Chrome add-on in this commit, because Manifest V2 is
already out of date and no one uses our Chrome add-on in actual.